+++ src/main/java/pe/edu/untels/certificadosdrsu/controllers/UsuarioAuditoriaController.java
@@ -39,7 +39,7 @@ public class UsuarioAuditoriaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> buscarPorId(@PathVariable int id) {
+    public ResponseEntity<?> buscarPorId(@PathVariable Long id) {
         ModelMapper m = new ModelMapper();
         Optional<UsuarioAuditoria> entidad = service.listId(id);
         
@@ -52,7 +52,7 @@ public class UsuarioAuditoriaController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> modificar(@PathVariable int id, @RequestBody UsuarioAuditoriaDTO dto) {
+    public ResponseEntity<?> modificar(@PathVariable Long id, @RequestBody UsuarioAuditoriaDTO dto) {
         ModelMapper m = new ModelMapper();
         Optional<UsuarioAuditoria> existente = service.listId(id);
         
@@ -68,7 +68,7 @@ public class UsuarioAuditoriaController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable int id) {
+    public ResponseEntity<?> eliminar(@PathVariable Long id) {
         Optional<UsuarioAuditoria> existente = service.listId(id);
         
         if (existente.isPresent()) {
