+++ src/main/java/pe/edu/untels/certificadosdrsu/entities/TipoProyecto.java
@@ -9,7 +9,7 @@ public class TipoProyecto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_tipo_proyecto")
-    private Integer idTipoProyecto;
+    private Long idTipoProyecto;
 
     @Column(name = "tipo_proyecto_nombre", nullable = false)
     private String tipoProyectoNombre;
@@ -18,16 +18,16 @@ public class TipoProyecto {
 
     }
 
-    public TipoProyecto(Integer idTipoProyecto, String tipoProyectoNombre) {
+    public TipoProyecto(Long idTipoProyecto, String tipoProyectoNombre) {
         this.idTipoProyecto = idTipoProyecto;
         this.tipoProyectoNombre = tipoProyectoNombre;
     }
 
-    public Integer getIdTipoProyecto() {
+    public Long getIdTipoProyecto() {
         return idTipoProyecto;
     }
 
-    public void setIdTipoProyecto(Integer idTipoProyecto) {
+    public void setIdTipoProyecto(Long idTipoProyecto) {
         this.idTipoProyecto = idTipoProyecto;
     }
