+++ src/main/java/pe/edu/untels/certificadosdrsu/repositories/IEnvioCorreoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.untels.certificadosdrsu.entities.EnvioCorreo;
 
 @Repository
-public interface IEnvioCorreoRepository extends JpaRepository<EnvioCorreo, Integer> {
+public interface IEnvioCorreoRepository extends JpaRepository<EnvioCorreo, Long> {
 }
