+++ src/main/java/pe/edu/untels/certificadosdrsu/servicesinterface/ICertificadoService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 public interface ICertificadoService {
     public List<Certificado> list();
     public Certificado insert(Certificado c);
-    public Optional<Certificado> listId(int id);
+    public Optional<Certificado> listId(Long id);
     public Certificado update(Certificado c);
-    public void delete(int id);
+    public void delete(Long id);
 }
